+++ cantera/trunk/test_problems/cathermo/HMW_graph_HvT/HMW_graph_HvT.cpp
@@ -23,7 +23,7 @@ int main(int argc, char** argv)
 {
 
     int retn = 0;
-    int i;
+    size_t i;
 
     try {
 
@@ -43,7 +43,7 @@ int main(int argc, char** argv)
         Cantera::ThermoPhase* solid = newPhase("NaCl_Solid.xml","NaCl(S)");
 
 
-        int nsp = HMW->nSpecies();
+        size_t nsp = HMW->nSpecies();
         double mf[100];
         double moll[100];
         for (i = 0; i < 100; i++) {
@@ -58,8 +58,8 @@ int main(int argc, char** argv)
 
         HMW->setState_TP(298.15, 1.01325E5);
 
-        int i1 = HMW->speciesIndex("Na+");
-        int i2 = HMW->speciesIndex("Cl-");
+        size_t i1 = HMW->speciesIndex("Na+");
+        size_t i2 = HMW->speciesIndex("Cl-");
         //int i3 = HMW->speciesIndex("H2O(L)");
         for (i = 1; i < nsp; i++) {
             moll[i] = 0.0;
@@ -206,20 +206,16 @@ int main(int argc, char** argv)
              * units of kJ/gmolSaltAdded
              */
             double phiL = L / Xmol[i1];
-
-
             double Aphi = HMW->A_Debye_TP() / 3.0;
             double AL = HMW->ADebye_L();
             double LrelMol = HMW->relative_molal_enthalpy() * 1.0E-6;
 
-
-
-            for (int k = 0; k < nsp; k++) {
+            for (size_t k = 0; k < nsp; k++) {
                 Enth0_RT[k] *= RT * 1.0E-6;
             }
 
             molarEnth0 = 0.0;
-            for (int k = 0; k < nsp; k++) {
+            for (size_t k = 0; k < nsp; k++) {
                 molarEnth0 += Xmol[k] * Enth0_RT[k];
             }
